# Backend/YEEEE/openapi_server/models/register_user.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util
from openapi_server.util import logger


class RegisterUser(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, name=None, email=None, password=None):  # noqa: E501
        """RegisterUser - a model defined in OpenAPI

        :param name: The name of this RegisterUser.  # noqa: E501
        :type name: str
        :param email: The email of this RegisterUser.  # noqa: E501
        :type email: str
        :param password: The password of this RegisterUser.  # noqa: E501
        :type password: str
        """
        self.openapi_types = {
            'name': str,
            'email': str,
            'password': str
        }

        self.attribute_map = {
            'name': 'name',
            'email': 'email',
            'password': 'password'
        }

        self._name = name
        self._email = email
        self._password = password

    @classmethod
    def from_dict(cls, dikt) -> 'RegisterUser':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RegisterUser of this RegisterUser.  # noqa: E501
        :rtype: RegisterUser
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this RegisterUser.


        :return: The name of this RegisterUser.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this RegisterUser.


        :param name: The name of this RegisterUser.
        :type name: str
        """
        if name is None:
            logger.error("RegisterUser: Versuch, name=None zu setzen")
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def email(self) -> str:
        """Gets the email of this RegisterUser.


        :return: The email of this RegisterUser.
        :rtype: str
        """
        return self._email

    @email.setter
    def email(self, email: str):
        """Sets the email of this RegisterUser.


        :param email: The email of this RegisterUser.
        :type email: str
        """
        if email is None:
            logger.error("RegisterUser: Versuch, email=None zu setzen")
            raise ValueError("Invalid value for `email`, must not be `None`")  # noqa: E501

        self._email = email

    @property
    def password(self) -> str:
        """Gets the password of this RegisterUser.


        :return: The password of this RegisterUser.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this RegisterUser.


        :param password: The password of this RegisterUser.
        :type password: str
        """
        if password is None:
            logger.error("RegisterUser: Versuch, password=None zu setzen")
            raise ValueError("Invalid value for `password`, must not be `None`")  # noqa: E501

        self._password = password