# Backend/YEEEE/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Kahuut API
  version: 1.1.0
  description: Eine API zur Verwaltung von Nutzer*innen, Themen und Spielen, inkl. Multiplayer-Features.
servers:
  - url: /

paths:
  /auth/admin/register:
    post:
      operationId: auth_admin_register_post
      description: |
        Registriert einen neuen Admin. Die E-Mail muss eindeutig sein.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAdmin'
        required: true
      responses:
        "201":
          description: Admin erfolgreich registriert
        "400":
          description: Ungültige Eingabe
      summary: Admin registrieren
      tags: [Auth]
      x-openapi-router-controller: openapi_server.controllers.auth_controller

  /auth/login:
    post:
      operationId: auth_login_post
      description: |
        Authentifiziert einen Admin oder User anhand von E-Mail und Passwort.
        Gibt bei Erfolg einen Token zurück.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        "200":
          description: Erfolgreicher Login (Token wird zurückgegeben)
        "401":
          description: Ungültige Anmeldedaten
      summary: Login für Admin/User
      tags: [Auth]
      x-openapi-router-controller: openapi_server.controllers.auth_controller

  /auth/user/register:
    post:
      operationId: auth_user_register_post
      description: |
        Registriert einen neuen User. Die E-Mail muss eindeutig sein.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUser'
        required: true
      responses:
        "201":
          description: User erfolgreich registriert
        "400":
          description: Ungültige Eingabe
      summary: User registrieren
      tags: [Auth]
      x-openapi-router-controller: openapi_server.controllers.auth_controller

  /themen:
    get:
      operationId: themen_get
      description: |
        Gibt eine Liste aller Themen zurück. Optional kann nach öffentlichen Themen gefiltert werden.
      parameters:
        - in: query
          name: public
          required: false
          schema:
            type: boolean
          description: Wenn true, werden nur öffentliche Themen zurückgegeben.
      responses:
        "200":
          description: Themenliste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Thema'
      summary: Themen abrufen (optional gefiltert nach public)
      tags: [Themen]
      x-openapi-router-controller: openapi_server.controllers.themen_controller

    post:
      operationId: themen_post
      description: |
        Legt ein neues Thema an. Nur für Admins möglich.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Thema'
        required: true
      responses:
        "201":
          description: Thema wurde erstellt
        "401":
          description: Nicht autorisiert
      security:
        - BearerAuth: []
      summary: Neues Thema anlegen (Admin)
      tags: [Themen]
      x-openapi-router-controller: openapi_server.controllers.themen_controller

  /fragen:
    post:
      operationId: fragen_post
      description: |
        Legt eine neue Frage zu einem Thema an. Nur für Admins möglich.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Frage'
        required: true
      responses:
        "201":
          description: Frage wurde erstellt
      summary: Neue Frage anlegen
      tags: [Fragen]
      x-openapi-router-controller: openapi_server.controllers.fragen_controller

  /antworten:
    post:
      operationId: antworten_post
      description: |
        Legt eine neue Antwort zu einer Frage an. Nur für Admins möglich.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Antwort'
        required: true
      responses:
        "201":
          description: Antwort wurde erstellt
      summary: Neue Antwort anlegen
      tags: [Antworten]
      x-openapi-router-controller: openapi_server.controllers.antworten_controller

  /mehrspieler/join:
    post:
      operationId: mehrspieler_join_post
      description: |
        Tritt einem Mehrspieler-Quiz über einen Code bei.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MehrspielerJoinPostRequest'
        required: true
      responses:
        "200":
          description: Erfolgreich beigetreten
        "400":
          description: Ungültige Eingabe
      summary: Mehrspieler-Quiz beitreten (per Code)
      tags: [Mehrspieler]
      x-openapi-router-controller: openapi_server.controllers.mehrspieler_controller

  /mehrspieler/start:
    post:
      operationId: mehrspieler_start_post
      description: |
        Startet ein neues Mehrspieler-Quiz für ein Thema. Nur für Admins möglich.
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MehrspielerStartPostRequest'
        required: true
      responses:
        "200":
          description: Mehrspieler-Quiz gestartet
        "401":
          description: Nicht autorisiert
        "500":
          description: Interner Serverfehler
      summary: Mehrspieler-Quiz starten (nur Admin)
      tags: [Mehrspieler]
      x-openapi-router-controller: openapi_server.controllers.mehrspieler_controller

components:
  schemas:
    RegisterAdmin:
      type: object
      required: [email, name, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    RegisterUser:
      type: object
      required: [email, name, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    Login:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    Thema:
      type: object
      required: [name, public]
      properties:
        name:
          type: string
        code:
          type: string
        public:
          type: boolean

    Frage:
      type: object
      required: [frage, fk_ID_Themen]
      properties:
        frage:
          type: string
        fk_ID_Themen:
          type: string
          format: uuid

    Antwort:
      type: object
      required: [antwort, fk_ID_Frage]
      properties:
        antwort:
          type: string
        richtig:
          type: boolean
        fk_ID_Frage:
          type: string
          format: uuid

    MehrspielerJoinPostRequest:
      type: object
      required: [code, userId]
      properties:
        code:
          type: string
        userId:
          type: string

    MehrspielerStartPostRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_BearerAuth